requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.9.0
